app/requirements.txt
streamlit
pandas
numpy
spacy
scikit-learn
torch
black
pinecone-client
boto3
requests
